requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu126

transformers==4.53.2
torch==2.11.0+cu126
torchvision==0.26.0+cu126
torchaudio==2.11.0+cu126
peft==0.18.1
bitsandbytes==0.48.1
sentencepiece==0.2.1
sacremoses==0.1.1
IndicTransToolkit==1.1.1
protobuf==7.34.1
accelerate==1.13.0
huggingface_hub==0.36.2
tqdm==4.67.3
numpy==2.4.3
pandas<3.0
scikit-learn==1.8.0
scipy==1.17.1
regex==2026.3.32
requests==2.33.0
safetensors==0.7.0
gradio==5.24.0
einops==0.8.1
streamlit>=1.30.0
llama-cpp-python==0.3.9
